+++ ext_emconf.php
@@ -8,10 +8,10 @@ $EM_CONF[$_EXTKEY] = [
     'author_email' => 'info@wacon.de',
     'author_company' => 'WACON Internet GmbH',
     'state' => 'stable',
-    'version' => '1.5.0',
+    'version' => '2.0.0',
     'constraints' => [
         'depends' => [
-            'typo3' => '13.4.0-13.4.99',
+            'typo3' => '13.4.0-14.1.99',
         ],
         'conflicts' => [
         ],